9️⃣update: change var type from short to int

+++ 3D/main.cpp
@@ -20,8 +20,8 @@ int main (int argc, char** argv) {
     read_input(filename, &M);
     //M.report();
 
-    short num_nodes = M.get_quantity(NUM_NODES);
-    short num_elements = M.get_quantity(NUM_ELEMENTS);
+    int num_nodes = M.get_quantity(NUM_NODES);
+    int num_elements = M.get_quantity(NUM_ELEMENTS);
     Matrix K(num_nodes,num_nodes), local_Ks[num_elements];
     Vector b(num_nodes),           local_bs[num_elements];
